// jobPortal/src/app/admin/model/job-post.ts
export interface JobPost {
  _id: String;
  jobId: String;
  jobTitle: String;
  postDate: String;
  role: String;
  responsibility: String;
  compName: String;
  exp: Number;
  salaryR: String;
  noPos: Number;
  location: String;
  skills: String;
  degree: String;
  compInfo: String;
  empType: String;
  indType: String;
  keyword: String;
  jobDesc: String;
}